Revert "Generate PDF at preparing, accepted, and rejected stage (#1365)"
This reverts commit 8c1e737bca82bf9977bc3475e02140642a0be580.
