+++ app/jobs/build_submission_bundle_job.rb
@@ -7,13 +7,6 @@ class BuildSubmissionBundleJob < ApplicationJob
       return
     end
 
-    begin
-      submission.generate_form_1040_pdf("preparing")
-    rescue
-      submission.transition_to!(:failed, error_message: "Could not generate PDF Form 1040.")
-      return
-    end
-
     response = SubmissionBundle.build(submission, documents: ["adv_ctc_irs1040"])
     if response.valid?
       submission.transition_to!(:queued)
