+++ app/models/efile_submission.rb
@@ -76,18 +76,6 @@ class EfileSubmission < ApplicationRecord
     address_service
   end
 
-  def generate_form_1040_pdf(status = "manual")
-    filename ||= "IRS 1040 - TY #{tax_return.year} - #{status.upcase}"
-
-    ClientPdfDocument.create_or_update(
-      output_file: AdvCtcIrs1040Pdf.new(self).output_file,
-      document_type: DocumentTypes::Form1040,
-      client: client,
-      filename: filename,
-      tax_return: tax_return
-    )
-  end
-
   private
 
   def generate_irs_submission_id(i = 0)
