+++ app/models/intake/ctc_intake.rb
@@ -342,4 +342,14 @@ class Intake::CtcIntake < Intake
   def filing_jointly?
     client.tax_returns.last.filing_status_married_filing_jointly?
   end
+
+  def rrc_eligible_filer_count
+    return primary_tin_type == "ssn" ? 1 : 0 if tax_return(2020).filing_status_single?
+
+    # if one spouse is a member of the armed forces, both qualify for benefits
+    return 2 if [primary_active_armed_forces, spouse_active_armed_forces].any?("yes")
+
+    # only filers with SSNs (valid for employment) are eligible for RRC
+    [primary_tin_type, spouse_tin_type].count { |tin_type| tin_type == "ssn" }
+  end
 end
