+++ app/models/tax_return.rb
@@ -88,26 +88,16 @@ class TaxReturn < ApplicationRecord
     outstanding_recovery_rebate_credit
   end
 
-  def rrc_eligible_filer_count
-    return intake.primary_tin_type == "ssn" ? 1 : 0 if filing_status_single?
-
-    # if one spouse is a member of the armed forces, both qualify for benefits
-    return 2 if [intake.primary_active_armed_forces, intake.spouse_active_armed_forces].any?("yes")
-
-    # only filers with SSNs (valid for employment) are eligible for RRC
-    [intake.primary_tin_type, intake.spouse_tin_type].count { |tin_type| tin_type == "ssn" }
-  end
-
   def expected_recovery_rebate_credit_one
     EconomicImpactPaymentOneCalculator.payment_due(
-      filer_count: rrc_eligible_filer_count,
+      filer_count: intake.rrc_eligible_filer_count,
       dependent_count: intake.dependents.count(&:eligible_for_eip1?)
     )
   end
 
   def expected_recovery_rebate_credit_two
     EconomicImpactPaymentTwoCalculator.payment_due(
-      filer_count: rrc_eligible_filer_count,
+      filer_count: intake.rrc_eligible_filer_count,
       dependent_count: intake.dependents.count(&:eligible_for_eip2?)
     )
   end
