+++ spec/factories/efile_submissions.rb
@@ -14,19 +14,19 @@
 #
 FactoryBot.define do
   factory :efile_submission do
-    transient do
-      tax_year { 2020 }
-      filing_status { "single" }
-      metadata {}
-    end
-    tax_return { create :tax_return, :ctc, year: tax_year, filing_status: filing_status }
+    tax_return
 
     trait :ctc do
+      transient do
+        tax_year { 2020 }
+        filing_status { "single" }
+        metadata {}
+      end
+      tax_return { create(:tax_return, :ctc, year: tax_year, filing_status: filing_status) }
       after :create do |submission|
         create :address, record: submission
       end
     end
-
 
     EfileSubmissionStateMachine.states.each do |state|
       trait state.to_sym do
