+++ spec/models/ctc_intake_spec.rb
@@ -237,4 +237,78 @@ describe Intake::CtcIntake do
       end
     end
   end
+
+  describe "#rrc_eligible_filer_count" do
+    context "when filing status is single" do
+      let(:intake) { create :ctc_intake, primary_tin_type: :itin, client: client }
+      let(:client) { create :client, tax_returns: [create(:tax_return, year: 2020, filing_status: :single)] }
+      context "when the primary is using an ITIN" do
+        it "filer_count is 0" do
+          expect(intake.rrc_eligible_filer_count).to eq 0
+        end
+      end
+
+      context "when the primary is using an SSN" do
+        let(:intake) { create :ctc_intake, primary_tin_type: :ssn, client: client }
+        let(:client) { create :client, tax_returns: [create(:tax_return, year: 2020, filing_status: :single)] }
+        it "filer count is 1" do
+          expect(intake.rrc_eligible_filer_count).to eq 1
+        end
+      end
+    end
+
+    context "when filing with a spouse" do
+      let(:client) { create :client, tax_returns: [create(:tax_return, year: 2020, filing_status: :married_filing_jointly)] }
+      let(:spouse_military) { "no" }
+      let(:primary_military) { "no" }
+      let(:primary_tin_type) { "itin" }
+      let(:spouse_tin_type) { "itin"}
+      let(:intake) do
+        create :ctc_intake,
+               client: client,
+               spouse_active_armed_forces: spouse_military,
+               primary_active_armed_forces: primary_military,
+               spouse_tin_type: spouse_tin_type,
+               primary_tin_type: primary_tin_type
+      end
+
+      context "when a spouse is part of the armed forces" do
+        let(:spouse_military) { "yes" }
+        it "has a filer count of 2" do
+          expect(intake.rrc_eligible_filer_count).to eq 2
+        end
+      end
+
+      context "when primary is part of the armed forces" do
+        let(:primary_military) { "yes" }
+
+        it "has a filer count of 2" do
+          expect(intake.rrc_eligible_filer_count).to eq 2
+        end
+      end
+
+      context "when primary is using an ssn and spouse is using ITIN" do
+        let(:primary_tin_type) { "ssn" }
+
+        it "has a filer count of one" do
+          expect(intake.rrc_eligible_filer_count).to eq 1
+        end
+      end
+
+      context "when primary is using an ITIN and spouse is using SSN" do
+        let(:spouse_tin_type) { "ssn" }
+
+        it "has a filer count of one" do
+          expect(intake.rrc_eligible_filer_count).to eq 1
+        end
+      end
+
+      context "when both are using ITIN" do
+        it "has a filer count of 0" do
+          expect(intake.rrc_eligible_filer_count).to eq 0
+        end
+      end
+    end
+  end
+
 end
