+++ spec/state_machines/efile_submission_state_machine_spec.rb
@@ -1,11 +1,6 @@
 require 'rails_helper'
 
 describe EfileSubmissionStateMachine do
-
-  before do
-    allow(ClientPdfDocument).to receive(:create_or_update)
-  end
-
   describe "after_transition" do
     context "to preparing" do
       let(:submission) { create(:efile_submission, :new) }
